Convert sardine.cli into a regular package
This is needed to work with having namespace packages disabled.
`sardine.cli.main` will still exist, but the preferred entrypoint
should be `sardine.cli` or `sardine-config`.
